refactor(api): split fronting-session.service.ts into services/fronting-session/ (api-w0lh)
