+++ .beans/api-w0lh--refactor-fronting-sessionservicets-666-loc-into-se.md
@@ -1,10 +1,11 @@
 ---
 # api-w0lh
 title: Refactor fronting-session.service.ts (666 LOC) into services/fronting-session/
-status: todo
+status: in-progress
 type: task
+priority: normal
 created_at: 2026-04-21T13:56:56Z
-updated_at: 2026-04-21T13:56:56Z
+updated_at: 2026-04-21T21:07:24Z
 parent: api-6l1q
 ---
 
@@ -16,12 +17,12 @@ Currently concentrates co-fronting lifecycle / overlap handling / switch trackin
 
 ## Scope
 
-- [ ] Create apps/api/src/services/fronting-session/ directory
-- [ ] Split into per-verb files with index.ts re-exporter so caller imports stay stable
-- [ ] Keep existing public exports identical
-- [ ] Preserve all existing tests; no coverage regression
-- [ ] Each resulting file ≤300 LOC; stretch target 200 LOC
-- [ ] Follow the conventions established by api-trlq
+- [x] Create apps/api/src/services/fronting-session/ directory
+- [x] Split into per-verb files (Option E — NO barrel; callers import directly from verb files)
+- [x] Keep existing public exports identical
+- [x] Preserve all existing tests; no coverage regression
+- [x] Each resulting file ≤300 LOC; stretch target 200 LOC (max 233 in queries.ts)
+- [x] Follow the conventions established by api-trlq (Option E flat verb-file layout)
 
 ## Out of scope
