+++ apps/api/src/__tests__/services/fronting-session.service.integration.test.ts
@@ -21,18 +21,22 @@ import {
   createFrontingComment as createComment,
   deleteFrontingComment,
 } from "../../services/fronting-comment.service.js";
+import { createFrontingSession } from "../../services/fronting-session/create.js";
 import {
   archiveFrontingSession,
-  createFrontingSession,
   deleteFrontingSession,
-  endFrontingSession,
+  restoreFrontingSession,
+} from "../../services/fronting-session/lifecycle.js";
+import {
   getActiveFronting,
   getFrontingSession,
   listFrontingSessions,
   parseFrontingSessionQuery,
-  restoreFrontingSession,
+} from "../../services/fronting-session/queries.js";
+import {
+  endFrontingSession,
   updateFrontingSession,
-} from "../../services/fronting-session.service.js";
+} from "../../services/fronting-session/update.js";
 import {
   assertApiError,
   genCustomFrontId,
