+++ apps/api/src/routes/fronting-sessions/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../lib/audit-writer.js";
 import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { archiveFrontingSession } from "../../services/fronting-session.service.js";
+import { archiveFrontingSession } from "../../services/fronting-session/lifecycle.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
