+++ apps/api/src/routes/fronting-sessions/end.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { endFrontingSession } from "../../services/fronting-session.service.js";
+import { endFrontingSession } from "../../services/fronting-session/update.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
