+++ apps/api/src/routes/fronting-sessions/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { getFrontingSession } from "../../services/fronting-session.service.js";
+import { getFrontingSession } from "../../services/fronting-session/queries.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
