+++ apps/api/src/routes/fronting-sessions/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { restoreFrontingSession } from "../../services/fronting-session.service.js";
+import { restoreFrontingSession } from "../../services/fronting-session/lifecycle.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
