+++ apps/api/src/routes/fronting/active.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { getActiveFronting } from "../../services/fronting-session.service.js";
+import { getActiveFronting } from "../../services/fronting-session/queries.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
